RelationshipLine Checkpoint
Had to call it for the night, wanted to make sure code was pushed so that it was safe
