+++ UMLEditor/UMLEditor/Classes/RelationshipLine.cs
@@ -1,0 +1,246 @@
+﻿using System.Collections.Generic;
+using Avalonia.Controls;
+using Avalonia.Controls.Shapes;
+using Avalonia;
+using System;
+using Avalonia.Media;
+using Avalonia.Collections;
+
+namespace UMLEditor.Classes;
+
+/// <summary>
+/// RelationshipLine class for use in the main window.
+/// </summary>
+public class RelationshipLine
+{
+    
+    private const double SymbolWidth = 15;
+    private const double SymbolHeight = 10;
+    private const int LineThickness = 2;
+    private readonly IBrush _brush = Brushes.CornflowerBlue;
+    
+    /// <summary>
+    /// SourceClass control for relationship line
+    /// </summary>
+    public UserControl SourceClass { get; private set; }
+    
+    /// <summary>
+    /// Destination class for relationship line
+    /// </summary>
+    public UserControl DestClass { get; private set; }
+    // Resharper disable once NotAccessedField.local
+    
+    /// <summary>
+    /// Relationship type for relationship line
+    /// </summary>
+    public string RelationshipType { get; private set; }
+    
+    /// <summary>
+    /// Symbol for relationship line.
+    /// </summary>
+    public Polyline Symbol { get; private set; } = null!;
+
+    /// <summary>
+    /// List of line segments for relationship line
+    /// </summary>
+    public List<Line> Segments { get; private set; } = null!;
+
+    /// <summary>
+    /// Default Ctor
+    /// </summary>
+    /// <param name="startCtrl">Source class control</param>
+    /// <param name="endCtrl">Dest class control</param>
+    /// <param name="relationshipType">Relationship type for line</param>
+    public RelationshipLine(UserControl startCtrl, UserControl endCtrl, string relationshipType)
+    {
+        SourceClass = startCtrl;
+        DestClass = endCtrl;
+        RelationshipType = relationshipType;
+    }
+
+    /// <summary>
+    /// Draw function for line
+    /// </summary>
+    public void Draw(Canvas myCanvas)
+    {
+            // Calculate lengths of controls
+            double startHalfWidth = SourceClass.Bounds.Width / 2;
+            double startHalfHeight = SourceClass.Bounds.Height / 2;
+            double endHalfWidth = DestClass.Bounds.Width / 2;
+            double endHalfHeight = DestClass.Bounds.Height / 2;
+            // Initialize points to middle of controls
+            Point start = new Point(
+                SourceClass.Bounds.X + startHalfWidth,
+                SourceClass.Bounds.Y + startHalfHeight);
+            Point end = new Point(
+                DestClass.Bounds.X + endHalfWidth,
+                DestClass.Bounds.Y + endHalfHeight);
+
+            // Set points to draw lines
+            Point midStart;
+            Point midEnd;
+            List<Point> diamondPoints;
+            List<Point> trianglePoints;
+            if (Math.Abs(start.X - end.X) > Math.Abs(start.Y - end.Y))
+            {
+                // Arrow is horizontal
+                midStart = new Point(Math.Abs(start.X + end.X) / 2, start.Y);
+                midEnd = new Point(Math.Abs(start.X + end.X) / 2, end.Y);
+                if (start.X < end.X)
+                {
+                    // Goes from left to right
+                    start = new Point(start.X + startHalfWidth, start.Y);
+                    end = new Point(end.X - endHalfWidth - (2 * SymbolWidth), end.Y);
+                    diamondPoints = new List<Point> { 
+                        end,
+                        new(end.X + SymbolWidth,end.Y - SymbolHeight),
+                        new(end.X + (2 * SymbolWidth),end.Y),
+                        new(end.X + SymbolWidth,end.Y + SymbolHeight),
+                        end };
+                    trianglePoints = new List<Point> { 
+                        new(end.X,end.Y - SymbolHeight),
+                        new(end.X + (2 * SymbolWidth),end.Y),
+                        new(end.X,end.Y + SymbolHeight),
+                        new(end.X,end.Y - SymbolHeight)
+                    };
+                }
+                else
+                {
+                    // Goes from right to left
+                    start = new Point(start.X - startHalfWidth, start.Y);
+                    end = new Point(end.X + endHalfWidth + (2 * SymbolWidth), end.Y);
+                    diamondPoints = new List<Point> { 
+                        end,
+                        new(end.X - SymbolWidth,end.Y - SymbolHeight),
+                        new(end.X - (2 * SymbolWidth),end.Y),
+                        new(end.X - SymbolWidth,end.Y + SymbolHeight),
+                        end };
+                    trianglePoints = new List<Point> { 
+                        new(end.X,end.Y - SymbolHeight),
+                        new(end.X - (2 * SymbolWidth),end.Y),
+                        new(end.X,end.Y + SymbolHeight),
+                        new(end.X,end.Y - SymbolHeight)
+                    };
+                }
+            }
+            else
+            {
+                // Arrow is vertical
+                midStart = new Point(start.X, Math.Abs(start.Y + end.Y) / 2);
+                midEnd = new Point(end.X, Math.Abs(start.Y + end.Y) / 2);
+                if (start.Y < end.Y)
+                {
+                    // Goes top to bottom
+                    start = new Point(start.X, start.Y + startHalfHeight);
+                    end = new Point(end.X, end.Y - endHalfHeight - (2 * SymbolWidth));
+                    diamondPoints = new List<Point>
+                    {
+                        end,
+                        new(end.X + SymbolHeight, end.Y + SymbolWidth),
+                        new(end.X, end.Y + (2 * SymbolWidth)),
+                        new(end.X - SymbolHeight, end.Y + SymbolWidth),
+                        end
+                    };
+                    trianglePoints = new List<Point>
+                    {
+                        new(end.X + SymbolHeight, end.Y),
+                        new(end.X, end.Y + (2 * SymbolWidth)),
+                        new(end.X - SymbolHeight, end.Y),
+                        new(end.X + SymbolHeight, end.Y)
+                    };
+                }
+                else
+                {
+                    // Goes bottom to top
+                    start = new Point(start.X, start.Y - startHalfHeight);
+                    end = new Point(end.X, end.Y + endHalfHeight + (2 * SymbolWidth));
+                    diamondPoints = new List<Point>
+                    {
+                        end,
+                        new(end.X + SymbolHeight, end.Y - SymbolWidth),
+                        new(end.X, end.Y - (2 * SymbolWidth)),
+                        new(end.X - SymbolHeight, end.Y - SymbolWidth),
+                        end
+                    };
+                    trianglePoints = new List<Point>
+                    {
+                        new(end.X + SymbolHeight, end.Y),
+                        new(end.X, end.Y - (2 * SymbolWidth)),
+                        new(end.X - SymbolHeight, end.Y),
+                        new(end.X + SymbolHeight, end.Y)
+                    };
+                }
+            }
+
+            Line l1 = CreateRelationshipLine(start, midStart);
+            Line l2 = CreateRelationshipLine(midStart, midEnd);
+            Line l3 = CreateRelationshipLine(midEnd, end);
+            
+            Segments.Add(l1);
+            Segments.Add(l2);
+            Segments.Add(l3);
+
+            // Add lines to the canvas
+            myCanvas.Children.Add(Segments[0]);
+            myCanvas.Children.Add(Segments[1]);
+            myCanvas.Children.Add(Segments[2]);
+
+            // Draw the relationship symbol based on provided type
+            switch (RelationshipType)
+            {
+                case "aggregation":
+                    myCanvas.Children.Add(CreateRelationshipSymbol(diamondPoints));
+                    break;
+                case "composition":
+                    Polyline polyline = CreateRelationshipSymbol(diamondPoints);
+                    polyline.Fill = _brush;
+                    myCanvas.Children.Add(polyline);
+                    break;
+                case "inheritance":
+                    Symbol = CreateRelationshipSymbol(trianglePoints);
+                    myCanvas.Children.Add(Symbol);
+                    break;
+                case "realization":
+                    Segments[0].StrokeDashArray = new AvaloniaList<double>(5, 3);
+                    Segments[1].StrokeDashArray = new AvaloniaList<double>(5, 3);
+                    Segments[2].StrokeDashArray = new AvaloniaList<double>(5, 3);
+                    Symbol = CreateRelationshipSymbol(trianglePoints);
+                    myCanvas.Children.Add(Symbol);
+                    break;
+            }
+    }
+    
+    /// <summary>
+    /// Draws the relationship symbol from the given list of points
+    /// </summary>
+    /// <param name="points">A list of points for the vertices to draw</param>
+    /// <returns>The new relationship symbol</returns>
+    private Polyline CreateRelationshipSymbol(List<Point> points)
+    {
+        Polyline polyline = new Polyline();
+        polyline.Name = "Polyline";
+        polyline.Points = points;
+        polyline.Stroke = _brush;
+        polyline.StrokeThickness = LineThickness;
+        return polyline;
+    }
+
+    /// <summary>
+    /// Creates a line from the given start to end points
+    /// </summary>
+    /// <param name="lineStart">Point to start at</param>
+    /// <param name="lineEnd">Point to end at</param>
+    /// <returns>The new line</returns>
+    private Line CreateRelationshipLine(Point lineStart, Point lineEnd)
+    {
+        Line l = new Line();
+        l.Name = "Line";
+        l.StartPoint = lineStart;
+        l.EndPoint = lineEnd;
+        l.Stroke = _brush;
+        l.StrokeThickness = LineThickness;
+        l.ZIndex = 10;
+        return l;
+    }
+}
+
